+++ gurobi/solve0.cpp
@@ -22,7 +22,7 @@ if (argc <= 3) {
 bool debug = false;
 std::string solver_name;
 std::string model_name;
-std::vector<int> data;
+std::vector<size_t> data;
 
 std::vector<std::string> args(argv + 1, argv + argc);
 
@@ -39,7 +39,7 @@ while (i<args.size()) {
         solver_name = args[i++];
         model_name = args[i++];
         while (i<args.size()) 
-            data.push_back(std::stoi(args[i++]));
+            data.push_back(std::stoul(args[i++]));
     }
 }
